+++ src/main/java/in/mohamedhalith/dao/EmployeeDAO.java
@@ -11,7 +11,6 @@ import in.mohamedhalith.exception.DBException;
 import in.mohamedhalith.exception.ValidationException;
 import in.mohamedhalith.model.Employee;
 import in.mohamedhalith.model.LeaveBalance;
-import in.mohamedhalith.model.LeaveRequest;
 import in.mohamedhalith.util.ConnectionUtil;
 
 public class EmployeeDAO {
@@ -79,64 +78,7 @@ public class EmployeeDAO {
 		return employee;
 	}
 
-	/**
-	 * This method is used to update the leave balance of employee after applying
-	 * for a leave request
-	 * 
-	 * @param employee
-	 * @param type
-	 * @param duration
-	 * @return boolean
-	 * @throws DBException
-	 * @throws ValidationException
-	 */
-	public boolean updateLeaveBalance(String keyword, int employeeId, LeaveRequest leaveRequest) throws DBException {
-		Connection connection = null;
-		PreparedStatement statement = null;
-		String leaveType = leaveRequest.getType().toLowerCase();
-		String query = null;
-		String apply = "apply";
-		String cancel = "cancel";
-		String reject = "reject";
-		try {
-			connection = ConnectionUtil.getConnection();
-			switch (leaveType) {
-			case "sickleave":
-				if (keyword.equalsIgnoreCase(apply)) {
-					query = "update employee_leavebalance set leave_balance = leave_balance - ? where employee_id = ? and type_of_leave = \'sickleave\'";
-				} else if (keyword.equalsIgnoreCase(cancel) || keyword.equalsIgnoreCase(reject)) {
-					query = "update employee_leavebalance set leave_balance = leave_balance + ? where employee_id = ? and type_of_leave = \'sickleave\'";
-				}
-				break;
-			case "casualleave":
-				if (keyword.equalsIgnoreCase(apply)) {
-					query = "update employee_leavebalance set leave_balance = leave_balance - ? where employee_id = ? and type_of_leave = \'casualleave\'";
-				} else if (keyword.equalsIgnoreCase(cancel) || keyword.equalsIgnoreCase(reject)) {
-					query = "update employee_leavebalance set leave_balance = leave_balance + ? where employee_id = ? and type_of_leave = \'casualleave\'";
-				}
-				break;
-			default:
-				if (keyword.equalsIgnoreCase(apply)) {
-					query = "update employee_leavebalance set leave_balance = leave_balance - ? where employee_id = ? and type_of_leave = \'earnedleave\'";
-				} else if (keyword.equalsIgnoreCase(cancel) || keyword.equalsIgnoreCase(reject)) {
-					query = "update employee_leavebalance set leave_balance = leave_balance + ? where employee_id = ? and type_of_leave = \'earnedleave\'";
-				}
-			}
-			statement = connection.prepareStatement(query);
-			statement.setInt(1, leaveRequest.getDuration());
-			statement.setInt(2, employeeId);
-			int row = statement.executeUpdate();
-			boolean isUpdated = false;
-			if (row == 1) {
-				isUpdated = true;
-			}
-			return isUpdated;
-		} catch (ClassNotFoundException | SQLException e) {
-			throw new DBException(e, "Cannot apply the leave request");
-		} finally {
-			ConnectionUtil.closeConnection(connection, statement);
-		}
-	}
+	
 
 	/**
 	 * This method is used to return a specific employee. Employee ID is obtained
